HPCC-19787 Send timestamped signature for Dali file requests
When engines send getPermissions to Dali, send a digital signature of the "username;timestamp"
string and the CDateTime containing that same timestamp. Dali will use that
CDateTime and user to rebuild the string, and call digisign verify() to ensure
the digital signature sent is valid. Additionally, ensure the request did not
hijack a signature by ensuring the timestamp in the signature was within a
configurable threshold. If deemed valid, LDAP authentication is unnecessary.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>