HPCC-18448 Change dropzone display on ECLWatch System Server page
1. Move the drop zone list into a folder because a drop zone may
contain many drop zone servers. A user may open the folder to see
those drop zone servers.
2. Clean computer name column. In the new drop zone setting, the
names for some drop zone servers are hard-coded to be 'ServerList'
and the names for other servers are blank.

This fix is in legacy xslt code which is still used by new UI.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>