HPCC-20124 Return ECLServer queue object in WsSMC.Activity
One ECLServer may listen multiple queues. In the existing WsSMC
Activity method, the queue status for those queues are read and
concatenated into one string. It is difficult for an ESP client
to parse. In this fix, each queue will be an object under an
ECLServer object. The queue object has queue name, status, and
status details.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>