HPCC-20193 Return ECLCCServer name in WsSMC.Activity if available
In the existing /Status/Servers data tree, all of ECLCCServers are
named as 'ECLCCserver'. WsSMC.Activity has to use the name +
IP/port as ECLCCServer name. In HPCC-20153, correct ECLCCServer
name is added to /Status/Servers using @cluster. WsSMC.Activity
should use the name as ECLCCServer name if it is available.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>