HPCC-20330 Stop aggregate stream to avoid potential deadlock
The aggreate stream of a global hash aggregate was not being
stopped. If the hash aggregate was stopped early (e.g. due to a
firstn upstream), it could leave unread data on the look ahead
pipe associated with the underlying pull distributor, which in
turn would cause the activity to hang when it was winding up.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>