HPCC-20557 Return special error for wrong esp session request
If a session enabled ESP was used and some web page is still open,
the session timer may trig a session lock request to ESP even if
the ESP has been switched to a session disabled ESP. In this fix,
ESP will detect the problem and return an 'Action not supported:...'
error to the session timer. Also clear session cookies from ESP
side.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>