HPCC-20437 Set limit of returned files in DropZoneFileSearch
The existing DropZoneFileSearch fails when too many files
match with a wildcard filter. In this fix, a max file limit
is introduced. If >1000 files match the search filter, only
1000 files are returned with a warning message. The code for
the file search is also simplified by passing the 'sub=true'
option. And a client has to specify a valid dropzone name, a
valid dropzone server, and a valid file name filter to do
the search.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>