HPCC-20712 Roll back code for authorizing user when restoring WU
This fix rolls back a code change for https://track.hpccsystems.com/
browse/HPCC-19807. Before HPCC-19807, WU Restoring code authorizes
user after the WU is restored. That is because the WU information is
needed for user authorization. If the user is not authorized, the
user cannot see the WU even when the WU is restored. Sasha server
will archive the WU automatically.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>