HPCC-20748 Skip wuid.dependencies.xml when listing archived WUs
The wuid.xml is using to list archived WUs. In the existing code,
the .xml is used as a key word to retrieve the wuid.xml files. It
does not work when wuid.dependencies.xml files also exist in the
same folder. In this fix, extra code is added to filter out
wuid.dependencies.xml files.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>