HPCC-20779 Configure dafilersv file access with default keys
The dafilesrv external file access service will only accept
signed requests, but the stock environment did not ship with a
suitable configuration.

Use the default generated keys in the stock environment for
dafilesrv cluster keys.
But also remove the default passphrase that these auto-generated
keys used to use since:
a) it is not helpful to keep poorly encoded passphrases in
widely distributed configuration files (environment.xml)
b) The dafilesrv signature configuration doesn't allow
passphrases to be configured in the environment (because of (a))

If passphrases are to be used, they should not be
stored/distributed in config files, but should be secured
elsewhere, e.g. using ssh-agent

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>