HPCC-20860 Add session request cookie to identify session request
If a request with an HTTP Authorization header is received, the
existing ESP code handles the request as non-session based request.
This is not correct because browser may remember old username/
password using the Authorization header and send the header with
session based request. In this fix, ESP/ECLWatch will set a cookie
when redirecting initial ECLWatch request to the login page. The
cookie is used to identify the following ECLWatch requests.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>