HPCC-20910 Prevent slave initialization aborting too early
Extend timeout for activity initialization time and make it
configurable, plus log why waiting in logs.
Prior to this change, if a query was blocked e.g. due to a file
locking issue in the master during initialization, it would cause
the slaves to fail after timeout (25 mins).
Not it will by default continue for up to 2hrs, configurable
via global/workunit option "actInitWaitTimeMins"

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>