HPCC-20116 Landing Zone Filter set not loading
The filter capabilities in the Landing Zone now returns a flat list and
will prepend any paths that a file belongs in as a result. There is no
expansion of directories any longer needed. This also fixes the bug of
directories not able to expand due to missing network address.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>

Signed-off-by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>