HPCC-21090 Extend default LDAP cache timeout
Reloading the LDAP security cache can take a while since there are quite
a few hpccinternal and other managed scopes. Currently the default for
a cache refresh is 5 minutes, which is way too often. This PR changes the
default for cache reload to every hour.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Reloading the LDAP security cache can take a while since there are quite
a few hpccinternal and other managed scopes. Currently the default for
a cache refresh is 5 minutes, which is way too often. This PR changes the
default for cache reload to every hour.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>