HPCC-20930 Potential crash in KJ with translation
When KJ handlers are handling multiple parts, the initialization
of key managers and their translators after the 1st part was
causing a RtlRecord to be released that was still in use by
the previously created key manager.
This could cause a crash.
Quite often it got away with it because assuming the parts
translators were the same (not necessarily in the case of super
keys), the RtlRecord released was reallocated and by luck
could occupy the same memory space.

This was spotted in a multi-channel setup where the chance
of the allocation being the same were far less.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>