HPCC-21059 Enhance custom request transform targeting
Support target xpath at the xsdl:choose level.  This allows the
choose to reposition the target position in the original xml that
all child setValue and appendValue actions will affect.

Also push the transform level target xpath down to it's child
xsdl:choose level before merging the transforms.  This means the
uniqueness of the target xpaths will be correct after the merge
is done.  Otherwise the transform level xpath can be over ridden
during the merge.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>