HPCC-21156 Fix problem with TLK and PARTITION keys
If PARTITION keys had 0 rows on some slaves, the TLK was not
being built correctly and as a consequence reading the key led
to the wrong answers coming back from getPartition() and lookups
targeted against the wrong parts, leading to missing matches.

Also avoid setting last value in TLK to max in local key case.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>