HPCC-19268 Logging in not unlocking other tabs
We provide the user the ability to lock one tab and lock others that
match that hostname/IP without interaction on each tab, the same goes with unlocking. When a user opens a
tab and they have other possible locked sessions the log in dialog is
presented. We now allow them to unlock other locked tabs in the same
fashion with a authenticated re-login.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>