HPCC-21008 Montitoring not handling mixed diff_monitor values correctly
esdl2monitor.xslt can generate the proper code but the wrong template
was being called.

The code in question deals with scalar fields in shared record
layouts that are monitored differently in different instances.
For example an address structure where the Line2 field is
monitored in some child datasets but not others.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>