HPCC-19155 Multiple problems with lazy caching.
When the limit is exceeded, it could cause indexes that used
that lazy cache to fail dealing with a file that had been closed.
Resulting in a checked_pread error (if local), or a
RFSERR_InvalidFileIOHandle if dealing with a remote file.

It could also cause a crash, if multiple uses of the same
physical file raced in and out of the cache and were then removed
after use, due to a bug in the handling of the objects during
removal from the cache.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>