HPCC-21306 Fix issue with PARTITION keys with 0 recs in part 0
When a PARTITION key is built, need to ensure each part has an
entry in the TLK. HPCC-21156 spotted/fixed the general case, but
part 0 is handled slightly differently and takes a different code
path, and was not addressed in HPCC-21156.
Spotted by a regression suite running in a 4(slaves)x4(ch) that
provoked 1st part of some built keys to have 0 records.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>