HPCC-21281 Sasha iterate format must contain column for all fields
HPCC-20036 introduced a bug, which meant if some of the requested
fields did not have entries, the field entry was skipped in the
returned result. This cause the client to misinterpret the
results, as the fields were not in the sequence it had requested.
The example where this problem was seen, was when esp requested
a archived workunit list and some workunits did not have a
jobName (a field esp had requested). Ultimately this caused a
assertex(SELF::isItem(pos)) to be hit.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>