HPCC-21391 ESP logging user's password
In CSoapService::processRequest(), the code queries a pointer to the context
userID .  In the call to processHeader(), that ctx pointer gets changed to point
to the password (or possibly something else). If a SOAPCALL error occurs, the logging
attempts to log the userID, but it now points to the pwd.
This PR only queries the name when needed

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexisrisk.com>