HPCC-21524 Check thor log restarting when retrieving WU thor log
If thor master/slave is restarted before a WU is finished, the
"Finished wuid=..." will not be in thor log. The existing code
for retrieving WU thor log only uses the "Finished wuid=..." to
find out the end of the WU thor log. In this fix, we also check
the thor log restarting (by checking log ID) for the end of the
WU thor log.

Also adding code to retrieve WU thor logs using a thread pool
for a better performance.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>