HPCC-21935 Fix lazy file io crash
It was possible for requests from the cache to overlap with
previous requests expiring and being removed.
The overlaping request for a previously cached result, that
was about to be destroyed linked the result before the
PSEUDO_COUNT was set, meaning it could be destroyed twice.

Protect by checking isAlive() when looking up and locking
mutex within beforeDispose()

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>