HPCC-22085 Fix connection table crit being held too long issues
A global mutex used when interacting with a table that caches
connections was being held for too long, in particular it was
being held when checking if the socket was stale or not. This
had the effect of causing all remote file socket connections who
hit a cache match at approx the same time (e.g. during a 400x
connect) to in effect be serialized, causing significant and
unnecessary slowdowns. On a 400x cluster, some operations could
take 15+ minutes before this change and ~15 seconds after.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>