HPCC-22209 Support ECLIDE in ESP usernameonly mode
When ESP is in usernameonly mode, the existing code always
checks the user name in HTTP cookie. If it is not found, ESP
redirects the request to a UserNameInput page. In this fixe,
ESP also checks the user name in HTTP Authorization header.
If not found, ESP will send 401 Unauthorized.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>