HPCC-22389 HPCC Admin group should allow fully qualified OU specification in config
Currrently, you can specify an LDAP group name in the configuration to
specify HPCC administrators. This group must be located under the "groups"
base dn which is also specified in the config. This PR allows you to
specify the group name as it does now, or a fully qualified OU that
resides elsewhere. With this new behaviour, the admin group membership
could be shared between HPCC deployments that do not share the same group
base DN.  This PR also fixes a misspelling the dali.xsd file

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexisrisk.com>