HPCC-13933 Keyed Join/Distribute not picking up replicate TLK
This was caused because these activities estalished the part
to read by 'getBestFilePart'. If getBestFilePart spotted that
the primary copy was missing an exception is fired.
If getBestFilePart is given a IExceptionHandler if passes
the exception on to it, if it doesn't it simply throws and
aborts.

In the Keyed Join/Distribute cases an exception handler was not
being passes and so getBestFilePart aborted prematurely, instead
of just warning about the missing primary and continuing to look
at other copies.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>