HPCC-22385 Close connection if upload file throws exception
When uploading a big file, Internet browser sends the file
to ESP block by block. If an exception (such as invalid
destination path) is found when a file block is received,
ESP should close socket connection to avoid more file blocks
from the browser. If the connection is not closed, the new
file block is treated as new ESP request. ESP may take a
weird action based on the content of the file block.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>