HPCC-22644 Fix KJ open part issue, if lots of parts off cluster.
If many of the parts are not local to the cluster the KJ is
running on, a maximum per handle limit is hit.
At that point it uses an existing handler.
However in the case of off-cluster handlers, it was incorrectly
using the 1st handler.
That would still work, if the 1st handler happened to be a
local/direct handler. However if it was a remote handler, the
requests would be sent to the remote slave and that service
would fail to open the part.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>