HPCC-22657 Fix unittest issues using uncontactable IPs
HPCC-22627 meant the IP's in a file being published, might
be contacted if the meta size info was missing.

In the case of some unit test logical files, the IPs
were made up, and no part size meta info was added.
That causes a hang trying to contact the made up IPs.

Fix by publilshing with size meta data for the purposes
of these unit tests.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>