HPCC-22114 Fix race condition crash handling grouped keyed join
The handling of the eog marker was wrong, and liable to cause
corruption. It set a eog flag into the 1st CJoinGroup of a
LHS group, but by the time it did it could have been freed.
It was also unnecessary since the processing of this flag
relied on seeing it in the start of the next group, so it
wasn't really an eog marker, but a group-start marker.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>