HPCC-22709 CDaliDownMonitor shouldn't be called on exit
The connection monitor designed to spot when Dali shutdown,
was also being called when the client (eclagent) stopped MP.
That could cause a hang (60 secs).

The monitor should only be called if/when Dali goes down.
To avoid it being called when the process stops MP, it needed
to be removed before that point.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>