HPCC-22911 Fix translate issue with uncompressed fixed width files
When reading a file with a record definition that has a fixed
record format that is also uncompressed, there were some checks
that validated that the physical file size was a multiple of the
fixed record size, and fired an error if not.
These checks are no longer valid, since dynamic record translation,
can translate between the two formats.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>