HPCC-22934 Ensure slave requesting act. init gets it.
A child query that requests activity initialization info. has
already determined if it requires it, i.e. that it either
requires a reinit (e.g. due to dynamic filename), or it is the
1st execution.
That is also the case if parallel child query graphs are
involved and they have yet to initialize.

However the bug fixed here, is that the master was ignoring
any requests for initialization data for activities that it
thought were already initialized.
This check was unecessary since the slave(s) had already checked,
and wrong in the case of child graphs cloned for parallel
execution.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>