HPCC-22240 Add auth_feature security declarations to dynamic bindings
--------------------------------------------------------------------------------
Recognize auth_feature attributes specified on Binding/Definition and
Binding/Definition/Methods/Method elements. The syntax used is the same as is
used in ESDL definitions. Where an ESDL definition recognizes EsdlService and
EsdlMethod as security fragment scopes, the binding supports BindingService and
BindingMethod. BindingMethod takes precedence over BindingService, which takes
precedence over EsdlMethod, which takes precedence over EsdlService.

Add defaultFeatureAuth to the DynamicESDL Esp Service configuration element. The
default value for this attribute, which will be applied if it is either missing
empty, is ${service}Access:full. In addition to supporting the auth_feature
notation, the value NO_DEFAULT may be used to require that security be declared
by either the ESDL definition, the dynamic binding, or a combination of both.

Signed-off-by: Tim Klemm <tim.klemm@lexisnexisrisk.com>