HPCC-23117 Embedded java not working properly in Roxie
A proper fix for the issues caused by thread-local object reuse by pooled
threads is beyond the scope of 7.2.x, so turn off the use of threadLocal in
Java for now.

That revealed some issues in the implementation of PERSIST('thread') when
threadLocal mode not available, so fix those too.

Also fix behaviour of PERSIST('channel') in Roxie for consistency.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>