HPCC-23130 Re-enable threadLocal option for embedded Java
Embedded Java's threadLocal option was disabled in 7.2.52 and upmerged
versions, because it was not working correctly (causing cores) if a thread in
a threadpool was reused (as is common in Roxie).

Correcting this so that threadLocal can be re-enabled requires a change to the
threadTermHook mechanism so that we can distinguish a threadpool end from a
thread about to be killed, and a codegen change so that the ICodeContext can
be passed per call rather than only at function context creation.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>