HPCC-23163 KJ fix for remote handling of parts from other clusters
If a smaller Thor was reading a file from a larget Thor, where
that larger Thor file overlapped, then some slaves on the smaller
cluster could find non-primary copies of the file local to it.
That meant other slaves would send requests to those slaves, but
they did not map to correct (non-primary) copy, and results in a
'Failed to open' file error.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>