HPCC-23035 Don't trace values written to workunit variables
Temporary values stored in Workunit variables may contain sensitive
information - there's no real benefit to outputting them in the tracing and it
can cause some grief.

Note that temporaries are deleted from the workunit at the end of the job and
so cannot be observed in the workunit XML either.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>