HPCC-22955 Add caching of public/private key files
To speed up access to key files, this PR
1) Defines an owned pub/priv key cache using unordered map
2) Declares and instance of that cache using Thread Local Storage
3) Calls the TLS based cache whenever these keys are required
Note that the cache will only exist for the duration of the current thread,
making it unlikely to be shared by other parallel graphs. Valgrind tested
to ensure no memory leaks. Also note that multiple key files can be shared within
the same graph, so a map is necessary over simply caching a single key

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexisrisk.com>