HPCC-23234 Fix problems if KJ in CQ and stopped early.
If a KJ is inside a CQ and it's stopped early (e.g. due to a
downstream CHOOSEN), then it can lead to spurious results,
because it is not reset fully on restart.

Also problems with the limiters not being reset, which effectively
meant that it had an unbound lookahead.
That could cause memory issues, probably more of an issue if
KJ is inside a LOOP where each execution would be expected to
create larger outputs than if inside a CQ.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>