HPCC-19274  Add ECLArchiveViewer
Add "ECLNew" tab that displays a list of ECL Archive sections on the left.

Clicking on a section name will display the contents on the right.

The ECL query is displayed on the right by default. If no archive contents are found for a WU the left side "DirectoryTree" is hidden.

Signed-off-by: Jaman Brundage <jbrundage372@gmail.com>