HPCC-23762 Thor linger
Allow thor to linger a configurable amount of time after it
has complete executing a graph.
The job/workflow detected that the workunit is associated with
a lingering Thor, it will communicate with it directly if it
can and avoid the latency of stopping/starting new master and
slave containers.

Also make a change to master/slave so that if containerPerAgent
is off in thor agent, then ensure master doesn't use a fix
port (akin to hthor process).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>