HPCC-23832 DESDL 'Roxie Test' form shows transformed request
To aid troubleshooting and QA, show the outgoing request to the Roxie after
it has been processed by Custom Request Transforms and legacy gateway
transformation. Previously the request was displayed before these
transformations were made.

The Roxie Test button visibility is controlled by an attribute on the ESDL
binding named 'includeRoxieTest'. To maintain backward compatibility the
button is visible by default.

A single function of the ESDLService object encapsulates all the steps
to transform the request.

Release exceptions caught in handleJSONPost to stop a leak.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>