HPCC-23984 Implement correct logic in CEspConfig::ensureSDSSessionDomains()
1. Rename the ensureSDSSessionDomains() to ensureSDSSession();
2. Inside the ensureSDSSession(), the code for reading Session
Domains settings is moved into new readSessionDomainsSetting().
It is called only esp is starting;
3. The ensureSDSSession() is called in 2 scenarios: 1. after
the readSessionDomainsSetting() is called and if esp is not detached
from dali; 2. when esp is attached to dali if the ensureSDSSession()
has not being called;
4. In the existing code, the EspHttpBinding::setSDSSession() is
used to ensure that /Sessions/Process has an Application branch
for the esp binding. It is called when EspHttpBinding is created.
That does not work if esp is not attached to dali when the
EspHttpBinding is created. In this PR, the related code is moved
into the ensureSDSSession() so that the code can also be called
when esp is attached to dali as needed.

Revise based on review
1. Change 1 MakeStringException to makeStringExceptionV;
2. Change 1 MakeStringException to makeStringException;
3. Change a while loop using for loop;
4. Add a const.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>