HPCC-24058 Add manifest command to esdl tool
The manifest command builds an ESDL service binding or bundle from an input
manifest file. The manifest contains the outline of the binding and can use
'Include' statements to load external files to include in the output. Supported
Include files are ESDL scripts, XSL transforms for logging and ESDL
definitions.

- Add tests to to exercise the esdl manifest command.
- Refactor the EsdlExample into separate independent examples, including one
  showing the use of the manifest tool.
- Add a README explaining the use of the new command.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>