HPCC-24117 Serial funnel deadlock fix
(The JIRA issue for example)
This issue occurred with a group funnel which always pulls its
inputs serially. It had a global SORT on one arm and a global
GROUP on the other.
The non-parallel version of funnel also started its inputs
sequentially.

Because some activities stop their input early (in start() if for
example it has exhausted it's local input, and because global SORT
depends on synchronizing stops across all slaves, the 1st input
can block on some slaves.
On others it can progress to the next funnel input, which can
block if not all slaves have started (e.g. in global GROUP).

One solution may be to ensure no activity ever stops its input
early in start(), though quite a few do, and I'm not sure that
would cover all cases of similar deadlock.

Another fix, employed here, is to ensure all funnel input start's
are started in parallel. This prevents activities like global
GROUP from deadlocking, and thus ensures progress so that the
global STOP will be stopped on all slaves.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>