HPCC-24241 Improve file check and ESP log in WsDfu DeleteFiles
1. The same file should not be deleted more than once even if
the same file name is in the request more than once. A clean
function is added to remove duplicated file names. The clean
function may be used by other ESP service methods later.
2. The existing ESP log has duplicated log lines for deleting
logical file. The existing WsDfu DeleteFiles deletes
requested files in two loops. In the first loop, the code
skips the non-super files. But, log lines 1371/1378 are not
skiped for those non-super files. This is fixed now.
3. Encapsulate the code into one class.

Revise based on review:
1. Use removeEntry() for now;
2. Add fileCanBeDeleted;
3. Release df before deleting the file;
4. Other small changes.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>