HPCC-24187 Parallel load roxie queries
Previous change was largely defeated by other global critical blocks within
the loading process.

This change reduces the blocked sections to just the adding of a new object to
the cache, and then loads the objects outside of the global critical block
(blocking only threads that were trying to load the same object).

Also renames some member variables (for clarity) and switches a
SpinLock to a CritSec.

Note that gcc does not handle exceptions in call_once code well, so we need to
do so manually.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>