HPCC-24599 Improve decoupled ESDL transaction logging
1. For decoupled logging, skip the loadPendingLogReqsFromExistingLogFiles()
call because CLogRequestReader is used to pickup jobs from tank files;
2. For decoupled logging, skip m_cleared (the file storing acked jobs)
related code because CLogRequestReader has its own logic to skip acked jobs;
3. Store the scriptValues into tank file so that the scriptValues can be
used by CLogRequestReader to skip some predefined jobs. The logging manager
creates a tank file using CLogFailSafe::Add() and CLogSerializer::Append().
Both CLogFailSafe::Add() and CLogSerializer::Append() are enhanced to
handle the scriptValues. The scriptValues input is set to nullptr where the
Append() and the Add() are not called for decoupled logging.
4. When retrieving a logging request from tank file, the CLogRequestReader
checks the scriptValues and skips the predefined jobs as needed.
5. Move the checkSkipThreadQueue() into the logging lib (logthread.cpp) so
that it can be used by both CLogRequestReader and logging manager.
6. Pass FailSafeLogsDir from logging manager to logging agent thread so
that the thread knows where to read tank files.
7. Do not set acked file if DecoupledLogging.
8. Pass the ScriptValues inside the filterLogContent.
9. Read tank files based on file name pattern.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>