HPCC-24666 Improve filterLogContent in ESDL logging
1. In the existing filterLogContent(), an extra tree layer is
added to the log request even if the log request is already
in a tree structure. That is fixed now. The related code is
moved in the front of the filterLogContent().

2. For the log request not in one tree structure (the log
request with ESPContext, UserContext, etc), only change in
this fix is to move the ScriptValues from the LogContent
branch to the UpdateLogRequest branch.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>