HPCC-23426 Add support for JSON style TxSummary
Add basic support for printing TxSummary as JSON, text or both. The output
style is configurable in the ESPProcess. Each element added to the summary
is associated with a group and log level. An element is printed only when
it matches the configured group and log level.

An element name can include JSON style dot path notation to indicate its
place in the output hierarchy for JSON style output. Element names with
dots output in the text style are printed as-is, they are not structured
in a hierarchy.

Standardize the timer interface to always require group and level which are
used to find matching timers. Timers with the same name only are not
considered a match and an exception is thrown.

The summary takes a profile to tailor its output just prior to serialization
in several ways:

 - It can rename entries for different output styles and groups.
 - A callback function can add entries or perform other steps as required.

Add additional entries in a separate 'enterprise' group for tracking
timings and values likely not of interest to existing ESP services.

Unittests are updated to exercise the new features.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>