HPCC-24969 Fix sasha archive age issue
The archiving mechanism checks to see if a workunit is old and
ready to be archived, based on the timestamps of recent activity,
record in the workunit.
This used to be recorded as TimeStamps/TimeStamp, but was changed
when the stats. was refactored (to Statistis/Statistic).
The sasha archiving code wasn't updated at that time.

The symptom of this bug, was that an old workunit, that was
recently active (e.g. a scheduled workunit that runs regularly),
could be archived when it should not have been.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>