HPCC-24950 Race condition on RoxieLocalState file
Since the cache is stored in a location that ends up shared between Roxie
nodes, it can clash on startup and lead to errors.

The purpose of the cache is to allow Roxie nodes that have been disconnected
from dali to restart. Moving the cache to local storage would be
pointless.

Something may need to take the place of this mechanism at some point, if we
are concerned about dali being a SPOF, but for now dieable it.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>