HPCC-24858 Allow dynamic logging manager configurations
Update the DESDL service to detect and use a logging manager defined in the
dynamic binding in addition to one defined in the static configuration. The
format of the dynamic configuration matches the format currently expected
from esp.xml.

A dynamically configured manager will always be used. A statically configured
manager will always be used in the absence of a dynamically configured
manager, meaning it persists in the presence of a dynamically configured
manager so it can take effect if the binding is republished without a
configured manager.

Signed-off-by: Tim Klemm <tim.klemm@lexisnexisrisk.com>