HPCC-25173 Fix 2 target related calls in cloud environment
1. Common out the code for validating target name.
   Both refreshValidClusters() and isValidCluster() are in ws_sql
service and ws_workunits service. The related code for those 2
methods are moved into the SMCLib. The refreshValidClusters() is
renamed to refreshValidTargets(). The isValidCluster() is renamed
as validateTargetName(). Both of them are upgraded to support both
bare metal environment and cloud environment. The validateTargetName()
is upgraded to thorw an exception if the target name is invalid.
The code in ws_sql service and ws_workunits service is modified
to match with those changes.
2. Add getThorClusterNames() which returns both thor target names
and thor queue names for both bare metal environment and cloud
environment. Replace the existing getEnvironmentThorClusterNames()
call in ws_smcService.cpp.
3. Simplify the code in the CWsSMCEx::onBrowseResources().

Revise based on review:
1. Change the MakeStringException() to makeStringExceptionV().
2. Change the refreshValidClusters() to static. Add targetsDirty
flag. Call the refreshValidClusters() if targetsDirty is true.
In cloud environment, only call refreshValidClusters() once.
3. Change the notEmpty() calls to isEmptyString().
4. Change the BoolHash validTargets to std::set<std::string> validTargets.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>