HPCC-25256 Prevent foreign subfile locking
When deleting a superfile it clears the SuperOwner meta data
from it's subfiles.
If some of those subfiles are foreign subfiles, it spuriously
attemps to create lock paths to the local env. file of the same
name (locking is not supported on foreign files).
As a consequence, it may block or be blocked, if the local named
file is locked or trying to lock.

Or, even worse (and how the bug was seen), if the subfile has
the same name as the parent, it will deadlock, as the parent is
already locked.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>