HPCC-25391 Streaming a PTree content to ESP response
In the existing ESP, when a PTree content is sent out as ESP
response content, the PTree has to be streamed into a file
and the file is streamed into the socket output of the ESP
Response. By this PR, ESP may support streaming a PTree content
directly into the socket.

Revised on based on review: optionally add xml header and xml stylesheet

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>