HPCC-25444 Redirect new ECLWatch homepage to login page
The existing code does not redirect new ECLWatch homepage to
the login page because the homepage /esp/files/index.html is
inside the /esp/files folder and most of the files inside the
folder are resources which do not need to be authenticated.
In this PR, the code is added to handle the new homepage
similar to old ECLWatch homepage /esp/files/stub.htm.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>