HPCC-25521 Change log request check in ESP Transaction log
The LogContentInFile request is used to let a logging agent know where
the original log request is in a tank file. For non-decouple logging,
ESP logging code has a function to check whether a log request is a
LogContentInFile request or original log request. The existing function
expects that both LogContentInFile request and original log requests
have the top xml tag which is incorrect for some original log requests.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>